# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-csc",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `csc(value)` by composing two float primitives.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified:\n\n1. ask `srvcs-sin` for `s = sin(value)`;\n2. ask `srvcs-floatdivide` for `result = 1 / s`.\n\nValidation is not handled here: this service never calls `srvcs-isnumber`\ndirectly. Its dependencies validate their own operands, and any `422` they\nraise is forwarded verbatim. If a dependency is unreachable it reports itself\ndegraded (`503`).",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CscResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CscResponse": {
        "type": "object",
        "required": [
          "value",
          "result"
        ],
        "properties": {
          "result": {
            "type": "number",
            "format": "double"
          },
          "value": {
            "type": "object"
          }
        }
      },
      "EvalRequest": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "object",
            "description": "The angle in radians to take the cosecant of."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}